Bring back old style 'ant pdf2' and add new 'ant hoss-pdf'
This restores SolrRefGuide.adoc and 'ant pdf2' to how cassandra had them (using jekylltest sources files)

Instead, this adds a new SolrRefGuide-all.adoc and uses it from a new 'ant hoss-pdf' which has all the new pdf generation
using all the adoc files from the automated dump with the 'ant build-nav' metadata.

The new PDF is also updated to not use the top level 'leveloffset' option, and to try and pull in the correct images
(although these still need to move someplace more sane).  Note: many of the images aren't currently fitting on the PDF
pages. (possibly becuase they are still using the 'inline' image syntax in the asciidoc files?)