fix the way leading/trailing whitespace is trimmed/moved from tags where it's generally not wanted
the previous code had an off by one bug in how it was selecting textnodes, which masked the fact that the rest of the code was very broken.

this commit generally improves on the overall situation, but as part of these changes i'm actaully leaving the whitespace in place in most cases since it's too easy to created non-intentional word concatenation otherwise.  Only in the case of paragraph tags do we explicitly trim.

This commit also fixes issue #20, which was the rabit hole i started down when i discovered this whole mess.